Improved naming consistency, version check
Version is now specified inside settings.gradle (rather than a VERSION file), and stored as a resource in qupath-core. This should help ensure the version is found even when running from an IDE with a different working directory.

The version check applied when attempting to read a project is relaxed, so as to handle mildly-corrupted projects (e.g. the ones I created when I accidentally wrote projects without having a version number available).